align aarch64, riscv64, loongarch64 stat structure padding type
this change is purely to document that they are the same in
preparation to remove the arch-specific headers for these archs and
replace them with a generic version that matches riscv32 and can be
shared by these and all future archs.

+++ arch/riscv64/bits/stat.h
@@ -6,7 +6,7 @@ struct stat {
 	uid_t st_uid;
 	gid_t st_gid;
 	dev_t st_rdev;
-	unsigned long __pad;
+	unsigned long long __pad;
 	off_t st_size;
 	blksize_t st_blksize;
 	int __pad2;
